try number 400

+++ OneOfEachStats1.java
@@ -27,7 +27,7 @@ public class OneOfEachStats1 {
 				sum++; //sums amount of kids in each trial					
 			}
 			
-			avg += (double) sum; //adds each sum to a total of kids in all trials
+			avg += sum; //adds each sum to a total of kids in all trials
 			if(sum == 2) fam2++; //adds to number of families with 2 kids
 			else if(sum == 3) fam3++; //adds to number of families with 3 kids
 			else fam4++; //adds to number of families with 4 or more kids
